Change benchmark order parameter normalisation
The Julia Optim library supports constraining a vector to unit norm.
In the relaxation benchmark, it makes sense to use this normalisation,
because the goal is to work with standard numerical libraries.

This doesn't make sense for the rest of my code, because it makes
the value of psi depend on the grid step.  My other code uses the
physics convention that psi is normalised to the number of atoms
in the condensate.